added the authMiddleware to the annonce methods & routes

+++ annonce-service/Controllers/AnnonceController.js
@@ -13,9 +13,11 @@ exports.createAnnonce = async (req, res) => {
       socketTimeoutMS: 45000
     });
 
-    const newAnnonce = new Annonce(req.body);
-    await newAnnonce.save();
-    res.status(201).json({ message: 'Annonce created successfully', newAnnonce });
+      const newAnnonce = new Annonce(req.body);
+      newAnnonce.id_vendeur = req.user._id;
+      console.log(req.user);
+      await newAnnonce.save();
+      res.status(201).json({ message: 'Annonce created successfully', newAnnonce });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -71,10 +73,20 @@ exports.updateAnnonceById = async (req, res) => {
       socketTimeoutMS: 45000
     });
 
-    const updatedAnnonce = await Annonce.findByIdAndUpdate(req.params.id, req.body, { new: true });
-    if (!updatedAnnonce) {
+    // Récupérer l'annonce à mettre à jour
+    const annonceToUpdate = await Annonce.findById(req.params.id);
+    if (!annonceToUpdate) {
       return res.status(404).json({ message: 'Annonce not found' });
     }
+
+    // Vérifier si l'utilisateur est autorisé à mettre à jour cette annonce
+    if (annonceToUpdate.id_vendeur.toString() !== req.user._id.toString()) {
+      return res.status(401).json({ message: 'You are Unauthorized to update this annonce' });
+    }
+
+    // Mettre à jour l'annonce
+    const updatedAnnonce = await Annonce.findByIdAndUpdate(req.params.id, req.body, { new: true });
+ 
     res.status(200).json({ message: 'Annonce updated successfully',updatedAnnonce});
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -91,12 +103,22 @@ exports.deleteAnnonceById = async (req, res) => {
       connectTimeoutMS: 30000,
       socketTimeoutMS: 45000
     });
-    
-    const deletedAnnonce = await Annonce.findByIdAndDelete(req.params.id);
-    if (!deletedAnnonce) {
+
+    // Récupérer l'annonce à supprimer
+    const annonceToDelete = await Annonce.findById(req.params.id);
+    if (!annonceToDelete) {
       return res.status(404).json({ message: 'Annonce not found' });
     }
-    res.status(200).json({ message: 'Annonce deleted successfully',deletedAnnonce});
+
+    // Vérifier si l'utilisateur est autorisé à supprimer cette annonce
+    if (annonceToDelete.id_vendeur.toString() !== req.user._id.toString()) {
+      return res.status(401).json({ message: 'You are Unauthorized to delete this annonce' });
+    }
+
+    // Supprimer l'annonce
+    const deletedAnnonce = await Annonce.findByIdAndDelete(req.params.id);
+    console.log(deletedAnnonce);
+    res.status(200).json({ message: 'Annonce deleted successfully', deletedAnnonce });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
